+++ R/igraph.r
@@ -9,7 +9,7 @@
 #' @aliases igraph.to.gexf gexf.to.igraph
 #' @param igraph.obj An object of class \code{igraph}.
 #' @param gexf.obj An object of class \code{gexf}.
-#' @param position A three-column data-frame with XYZ coords.
+#' @param ... Further arguments passed to \code{\link{gexf}}.
 #' @return \itemize{ \item For \code{igraph.to.gexf} : \code{gexf} class object
 #' \item For \code{gexf.to.igraph} : \code{igraph} class object }
 #' @author George Vega Yon \email{g.vegayon@gmail.com}
@@ -26,9 +26,17 @@
 #'   gexf1 <- read.gexf(fn)
 #'   igraph1 <- gexf.to.igraph(gexf1)
 #'   gexf2 <- igraph.to.gexf(igraph1)
+#'   
+#'   # Now, let's do it with a layout!
+#'   pos <- igraph::layout_nicely(igraph1)
+#'   plot(
+#'     igraph.to.gexf(igraph1, nodesVizAtt = list(position=cbind(pos, 0))),
+#'     edgeWidthFactor = .01)
 #'  }
 #' @export
-igraph.to.gexf <- function(igraph.obj, position=NULL) {
+igraph.to.gexf <- function(igraph.obj, ...) {
+  
+  dots <- list(...)
   
   g <- igraph.obj
   rm(igraph.obj)
@@ -77,56 +85,58 @@ igraph.to.gexf <- function(igraph.obj, position=NULL) {
   # Nodes Attributes
   x <- igraph::list.vertex.attributes(g)
   x <- x[!(x %in% c("label","color","size"))]
-  if (!length(x)) nAtt <- NULL
-  else nAtt <- subset(nodes, select=x)
+  if (length(x))
+    dots$nodesAtt <- subset(nodes, select=x)
   
   # Edges Attributes
   x <- igraph::list.edge.attributes(g)
   x <- x[!(x %in% c("weight","color","edgesLabel","width"))]
-  if (!length(x)) eAtt <- NULL
-  else eAtt <- subset(tmpedges, select=x)
+  if (length(x)) 
+    dots$edgesAtt <- subset(tmpedges, select=x)
   
   # Edges Weights
-  if (length(igraph::E(g)$weight)) eW <- igraph::E(g)$weight
-  else eW <- NULL
+  if (!length(dots$edgesWeight) && length(igraph::E(g)$weight))
+    dots$edgesWeight <- igraph::E(g)$weight
+  else
+    dots$edgesWeight <- NULL
   
   # Nodes Viz att
-  if (length(tmpnodes$color)) {
-    nVizAtt <- list(color=t(grDevices::col2rgb(tmpnodes$color, alpha=T)))
-    nVizAtt$color[,4] <- 1
-  }
-  else nVizAtt <- NULL
-  
-  if (length(tmpnodes$size)){
-    nVizAtt$size <- tmpnodes$size
+  if (!length(dots$nodesVizAtt$color)) {
+    if (length(tmpnodes$color)) {
+      dots$nodesVizAtt$color <- list(color=t(grDevices::col2rgb(tmpnodes$color, alpha=T)))
+      dots$nodesVizAtt$color[,4] <- 1
+    } else 
+      dots$nodesVizAtt$color <- NULL
   }
   
-  nVizAtt$position <- position
+  
+  if (length(dots$nodesVizAtt$size) && length(tmpnodes$size)){
+    dots$nodesVizAtt$size <- tmpnodes$size
+  }
   
   # Edges Viz att
-  if (length(tmpedges$color)) {
-    eVizAtt <- list(color=t(grDevices::col2rgb(tmpedges$color, alpha=T)))
-    eVizAtt$color[,4] <- 1 
+  if (!length(dots$edgexVizAtt$color)) {
+    if (length(tmpedges$color)) {
+      dots$edgexVizAtt$color <- list(color=t(grDevices::col2rgb(tmpedges$color, alpha=T)))
+      dots$edgexVizAtt$color[,4] <- 1 
+    } else
+      dots$edgexVizAtt$color <- NULL
   }
-  else eVizAtt <- NULL
   
   # Edge type
-  if (igraph::is.directed(g)) defaultedgetype <- "directed"
-  else defaultedgetype <- "undirected"
+  if (!length(dots$defaultedgetype) && igraph::is.directed(g)) 
+    dots$defaultedgetype <- "directed"
+  else 
+    dots$defaultedgetype <- "undirected"
   
   # Building graph
-  return(
-    gexf(
-      nodes = nodes[,c("id","name")], 
-      edges = edges, 
-      edgesAtt = eAtt,
-      nodesAtt = nAtt,
-      defaultedgetype = defaultedgetype,
-      nodesVizAtt = nVizAtt,
-      edgesVizAtt = eVizAtt,
-      edgesWeight = eW
-      )
-    )
+  do.call(gexf, c(
+    list(
+      nodes = nodes[,c("id","name")],
+      edges = edges
+      ), dots)
+  )
+
 }
 
 #' @export
@@ -138,10 +148,6 @@ gexf.to.igraph <- function(gexf.obj) {
  
   g <- gexf.obj
   rm(gexf.obj)
-  #colnames(g$nodes)[colnames(g$nodes)=="label"] <- "name"
-  
-  # Starting igraph object
-  #colnames(g$nodes)[colnames(g$nodes) == "name"]
   
   g2 <- igraph::graph.data.frame(
     g$edges[,c("source","target")],